refactor(converter): image-converter を lib/ に移動し instructions を刷新
- knowledge/converter.ts → lib/image-converter.ts に移動（汎用化）
- converter-agent の instructions を文書/写真両対応のプロンプトに更新
- モデルを GEMINI_PRO → GEMINI_FLASH に変更（速度・コスト改善）
