+++ server/src/services/knowledge/upload.ts
@@ -1,5 +1,5 @@
+import { convertToMarkdown, isSupportedMimeType } from "~/lib/image-converter";
 import { logger } from "~/lib/logger";
-import { convertToMarkdown, isSupportedMimeType } from "./converter";
 import { syncFile } from "./sync";
 
 const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB for Markdown
